Refactor code removing decodeCertificate in favor of the new one
